Massive refactoring of crates, model and snaphot.html.
Removed seperate click methods for "module" table and other tables.
Modified model to allow each table to add configuration as to which entries in each row can be clicked on to form URL elements for the RESTful web API.  - also defined which element is the "row id" of the table. If no indices are stored for defining the URL elements it just uses the row-ID.

This meant have to remove the old method of highlighting the element that was clicked on by linking to the DOM element and using binding to set its class. Prefer this way as well, seemed clumsy to have a reference to a DOM element in the model. Need to come up with another way of highlighting the clicked elements though. Could use a method that returns a class name.